requirements.txt
Flask
Flask-Mail
Pillow
gunicorn